Move CoverInfo to coverart.h. Add CoverArt struct.
